@@ -33,7 +33,6 @@ classifiers =
 
   Programming Language :: Python
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
@@ -43,7 +42,7 @@ classifiers =
   Topic :: Internet :: WWW/HTTP
 
 [options]
-python_requires = >=3.8
+python_requires = >=3.9
 packages = aiohttp
 # https://setuptools.readthedocs.io/en/latest/setuptools.html#setting-the-zip-safe-flag
 zip_safe = False
@@ -55,7 +54,7 @@ install_requires =
   async-timeout >= 4.0, < 5.0 ; python_version < "3.11"
   frozenlist >= 1.1.1
   multidict >=4.5, < 7.0
-  yarl >= 1.0, < 2.0
+  yarl >= 1.12.0, < 2.0
 
 [options.exclude_package_data]
 * =
@@ -87,9 +86,14 @@ max-line-length=79
 zip_ok = false
 
 [flake8]
-extend-select = B950
+extend-select =
+  B950,
+  # NIC001 -- "Implicitly concatenated str literals on one line"
+  NIC001,
+  # NIC101 -- "Implicitly concatenated bytes literals on one line"
+  NIC101,
 # TODO: don't disable D*, fix up issues instead
-ignore = N801,N802,N803,E203,E226,E305,W504,E252,E301,E302,E501,E704,W503,W504,D1,D4
+ignore = N801,N802,N803,NIC002,NIC102,E203,E226,E305,W504,E252,E301,E302,E501,E704,W503,W504,D1,D4
 max-line-length = 88
 per-file-ignores =
     # I900: Shouldn't appear in requirements for examples.
@@ -138,8 +142,6 @@ addopts =
 filterwarnings =
     error
     ignore:module 'ssl' has no attribute 'OP_NO_COMPRESSION'. The Python interpreter is compiled against OpenSSL < 1.0.0. Ref. https.//docs.python.org/3/library/ssl.html#ssl.OP_NO_COMPRESSION:UserWarning
-    ignore:unclosed transport <asyncio.sslproto._SSLProtocolTransport object.*:ResourceWarning
-    ignore:unclosed transport <_ProactorSocketTransport closing fd=-1>:ResourceWarning
     ignore:Unclosed client session <aiohttp.client.ClientSession object at 0x:ResourceWarning
     ignore:The loop argument is deprecated:DeprecationWarning:asyncio
     ignore:Creating a LegacyVersion has been deprecated and will be removed in the next major release:DeprecationWarning::
@@ -167,7 +169,6 @@ junit_suite_name = aiohttp_test_suite
 norecursedirs = dist docs build .tox .eggs
 minversion = 3.8.2
 testpaths = tests/
-junit_family=xunit2
 xfail_strict = true
 markers =
     dev_mode: mark test to run in dev mode.
