@@ -14,12 +14,16 @@ from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
+    Dict,
+    Generic,
     Iterator,
     List,
     Optional,
     Type,
+    TypeVar,
     Union,
     cast,
+    overload,
 )
 from unittest import IsolatedAsyncioTestCase, mock
 
@@ -35,17 +39,20 @@ from aiohttp.client import (
 )
 
 from . import ClientSession, hdrs
-from .abc import AbstractCookieJar
+from .abc import AbstractCookieJar, AbstractStreamWriter
 from .client_reqrep import ClientResponse
 from .client_ws import ClientWebSocketResponse
 from .helpers import sentinel
 from .http import HttpVersion, RawRequestMessage
-from .typedefs import StrOrURL
+from .streams import EMPTY_PAYLOAD, StreamReader
+from .typedefs import LooseHeaders, StrOrURL
 from .web import (
     Application,
     AppRunner,
+    BaseRequest,
     BaseRunner,
     Request,
+    RequestHandler,
     Server,
     ServerRunner,
     SockSite,
@@ -60,6 +67,14 @@ else:
 
 if sys.version_info >= (3, 11) and TYPE_CHECKING:
     from typing import Unpack
+
+if sys.version_info >= (3, 11):
+    from typing import Self
+else:
+    Self = Any
+
+_ApplicationNone = TypeVar("_ApplicationNone", Application, None)
+_Request = TypeVar("_Request", bound=BaseRequest)
 
 REUSE_ADDRESS = os.name == "posix" and sys.platform != "cygwin"
 
@@ -90,7 +105,7 @@ def unused_port() -> int:
         return cast(int, s.getsockname()[1])
 
 
-class BaseTestServer(ABC):
+class BaseTestServer(ABC, Generic[_Request]):
     __test__ = False
 
     def __init__(
@@ -105,10 +120,10 @@ class BaseTestServer(ABC):
         ] = get_port_socket,
         **kwargs: Any,
     ) -> None:
-        self.runner: Optional[BaseRunner] = None
+        self.runner: Optional[BaseRunner[_Request]] = None
         self._root: Optional[URL] = None
         self.host = host
-        self.port = port
+        self.port = port or 0
         self._closed = False
         self.scheme = scheme
         self.skip_url_asserts = skip_url_asserts
@@ -120,8 +135,6 @@ class BaseTestServer(ABC):
         self._ssl = kwargs.pop("ssl", None)
         self.runner = await self._make_runner(handler_cancellation=True, **kwargs)
         await self.runner.setup()
-        if not self.port:
-            self.port = 0
         absolute_host = self.host
         try:
             version = ipaddress.ip_address(self.host).version
@@ -144,14 +157,14 @@ class BaseTestServer(ABC):
         self._root = URL(f"{self.scheme}://{absolute_host}:{self.port}")
 
     @abstractmethod  # pragma: no cover
-    async def _make_runner(self, **kwargs: Any) -> BaseRunner:
+    async def _make_runner(self, **kwargs: Any) -> BaseRunner[_Request]:
         pass
 
     def make_url(self, path: StrOrURL) -> URL:
         assert self._root is not None
         url = URL(path)
         if not self.skip_url_asserts:
-            assert not url.is_absolute()
+            assert not url.absolute
             return self._root.join(url)
         else:
             return URL(str(self._root) + str(path))
@@ -165,7 +178,7 @@ class BaseTestServer(ABC):
         return self._closed
 
     @property
-    def handler(self) -> Server:
+    def handler(self) -> Server[_Request]:
         # for backward compatibility
         # web.Server instance
         runner = self.runner
@@ -189,10 +202,10 @@ class BaseTestServer(ABC):
             assert self.runner is not None
             await self.runner.cleanup()
             self._root = None
-            self.port = None
+            self.port = 0
             self._closed = True
 
-    async def __aenter__(self) -> "BaseTestServer":
+    async def __aenter__(self) -> Self:
         await self.start_server()
         return self
 
@@ -205,7 +218,7 @@ class BaseTestServer(ABC):
         await self.close()
 
 
-class TestServer(BaseTestServer):
+class TestServer(BaseTestServer[Request]):
     def __init__(
         self,
         app: Application,
@@ -218,14 +231,14 @@ class TestServer(BaseTestServer):
         self.app = app
         super().__init__(scheme=scheme, host=host, port=port, **kwargs)
 
-    async def _make_runner(self, **kwargs: Any) -> BaseRunner:
+    async def _make_runner(self, **kwargs: Any) -> AppRunner:
         return AppRunner(self.app, **kwargs)
 
 
-class RawTestServer(BaseTestServer):
+class RawTestServer(BaseTestServer[BaseRequest]):
     def __init__(
         self,
-        handler: _RequestHandler,
+        handler: _RequestHandler[BaseRequest],
         *,
         scheme: str = "",
         host: str = "127.0.0.1",
@@ -240,7 +253,7 @@ class RawTestServer(BaseTestServer):
         return ServerRunner(srv, **kwargs)
 
 
-class TestClient:
+class TestClient(Generic[_Request, _ApplicationNone]):
     """
     A test client implementation.
 
@@ -250,16 +263,32 @@ class TestClient:
 
     __test__ = False
 
+    @overload
     def __init__(
+        self: "TestClient[Request, Application]",
+        server: TestServer,
+        *,
+        cookie_jar: Optional[AbstractCookieJar] = None,
+        **kwargs: Any,
+    ) -> None: ...
+    @overload
+    def __init__(
+        self: "TestClient[_Request, None]",
+        server: BaseTestServer[_Request],
+        *,
+        cookie_jar: Optional[AbstractCookieJar] = None,
+        **kwargs: Any,
+    ) -> None: ...
+    def __init__(  # type: ignore[misc]
         self,
-        server: BaseTestServer,
+        server: BaseTestServer[_Request],
         *,
         cookie_jar: Optional[AbstractCookieJar] = None,
         **kwargs: Any,
     ) -> None:
         if not isinstance(server, BaseTestServer):
             raise TypeError(
-                "server must be TestServer " "instance, found type: %r" % type(server)
+                "server must be TestServer instance, found type: %r" % type(server)
             )
         self._server = server
         if cookie_jar is None:
@@ -281,16 +310,16 @@ class TestClient:
         return self._server.host
 
     @property
-    def port(self) -> Optional[int]:
+    def port(self) -> int:
         return self._server.port
 
     @property
-    def server(self) -> BaseTestServer:
+    def server(self) -> BaseTestServer[_Request]:
         return self._server
 
     @property
-    def app(self) -> Optional[Application]:
-        return cast(Optional[Application], getattr(self._server, "app", None))
+    def app(self) -> _ApplicationNone:
+        return getattr(self._server, "app", None)  # type: ignore[return-value]
 
     @property
     def session(self) -> ClientSession:
@@ -446,7 +475,7 @@ class TestClient:
             await self._server.close()
             self._closed = True
 
-    async def __aenter__(self) -> "TestClient":
+    async def __aenter__(self) -> Self:
         await self.start_server()
         return self
 
@@ -494,7 +523,7 @@ class AioHTTPTestCase(IsolatedAsyncioTestCase, ABC):
         """Return a TestServer instance."""
         return TestServer(app)
 
-    async def get_client(self, server: TestServer) -> TestClient:
+    async def get_client(self, server: TestServer) -> TestClient[Request, Application]:
         """Return a TestClient instance."""
         return TestClient(server)
 
@@ -575,16 +604,16 @@ def _create_transport(sslcontext: Optional[SSLContext] = None) -> mock.Mock:
 def make_mocked_request(
     method: str,
     path: str,
-    headers: Any = None,
+    headers: Optional[LooseHeaders] = None,
     *,
-    match_info: Any = sentinel,
+    match_info: Optional[Dict[str, str]] = None,
     version: HttpVersion = HttpVersion(1, 1),
     closing: bool = False,
-    app: Any = None,
-    writer: Any = sentinel,
-    protocol: Any = sentinel,
-    transport: Any = sentinel,
-    payload: Any = sentinel,
+    app: Optional[Application] = None,
+    writer: Optional[AbstractStreamWriter] = None,
+    protocol: Optional[RequestHandler[Request]] = None,
+    transport: Optional[asyncio.Transport] = None,
+    payload: StreamReader = EMPTY_PAYLOAD,
     sslcontext: Optional[SSLContext] = None,
     client_max_size: int = 1024**2,
     loop: Any = ...,
@@ -635,14 +664,14 @@ def make_mocked_request(
     if app is None:
         app = _create_app_mock()
 
-    if transport is sentinel:
+    if transport is None:
         transport = _create_transport(sslcontext)
 
-    if protocol is sentinel:
+    if protocol is None:
         protocol = mock.Mock()
         protocol.transport = transport
 
-    if writer is sentinel:
+    if writer is None:
         writer = mock.Mock()
         writer.write_headers = make_mocked_coro(None)
         writer.write = make_mocked_coro(None)
@@ -651,17 +680,13 @@ def make_mocked_request(
         writer.transport = transport
 
     protocol.transport = transport
-    protocol.writer = writer
-
-    if payload is sentinel:
-        payload = mock.Mock()
 
     req = Request(
         message, payload, protocol, writer, task, loop, client_max_size=client_max_size
     )
 
     match_info = UrlMappingMatchInfo(
-        {} if match_info is sentinel else match_info, mock.Mock()
+        {} if match_info is None else match_info, mock.Mock()
     )
     match_info.add_app(app)
     req._match_info = match_info
