@@ -115,7 +115,12 @@ def normalize_path_middleware(
 
 def _fix_request_current_app(app: "Application") -> Middleware:
     async def impl(request: Request, handler: Handler) -> StreamResponse:
-        with request.match_info.set_current_app(app):
+        match_info = request.match_info
+        prev = match_info.current_app
+        match_info.current_app = app
+        try:
             return await handler(request)
+        finally:
+            match_info.current_app = prev
 
     return impl
