@@ -9,6 +9,7 @@ from yarl import URL
 
 import aiohttp
 from aiohttp.client_reqrep import ClientRequest, ClientResponse
+from aiohttp.connector import _SSL_CONTEXT_VERIFIED
 from aiohttp.helpers import TimerNoop
 from aiohttp.test_utils import make_mocked_coro
 
@@ -94,6 +95,7 @@ class TestProxy(unittest.TestCase):
                 )
 
                 conn.close()
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -157,6 +159,7 @@ class TestProxy(unittest.TestCase):
                 )
 
                 conn.close()
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch(
         "aiohttp.connector.aiohappyeyeballs.start_connection",
@@ -206,6 +209,7 @@ class TestProxy(unittest.TestCase):
                 )
             self.assertEqual(req.url.path, "/")
             self.assertEqual(dict(req.headers), expected_headers)
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch(
         "aiohttp.connector.aiohappyeyeballs.start_connection",
@@ -244,6 +248,7 @@ class TestProxy(unittest.TestCase):
                     self.loop.run_until_complete(
                         connector.connect(req, [], aiohttp.ClientTimeout())
                     )
+        self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -324,6 +329,7 @@ class TestProxy(unittest.TestCase):
                             self.loop.run_until_complete(proxy_req.close())
                             proxy_resp.close()
                             self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -407,6 +413,7 @@ class TestProxy(unittest.TestCase):
                             self.loop.run_until_complete(proxy_req.close())
                             proxy_resp.close()
                             self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -488,6 +495,7 @@ class TestProxy(unittest.TestCase):
                             self.loop.run_until_complete(proxy_req.close())
                             proxy_resp.close()
                             self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -564,6 +572,7 @@ class TestProxy(unittest.TestCase):
                                         req, [], aiohttp.ClientTimeout()
                                     )
                                 )
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -638,6 +647,7 @@ class TestProxy(unittest.TestCase):
                                         req, [], aiohttp.ClientTimeout()
                                     )
                                 )
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -713,6 +723,7 @@ class TestProxy(unittest.TestCase):
                         self.loop.run_until_complete(proxy_req.close())
                         proxy_resp.close()
                         self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -782,6 +793,7 @@ class TestProxy(unittest.TestCase):
                                     req, [], aiohttp.ClientTimeout()
                                 )
                             )
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -828,6 +840,7 @@ class TestProxy(unittest.TestCase):
                     connector._create_connection(req, [], aiohttp.ClientTimeout())
                 )
                 self.assertEqual(req.url, URL("http://localhost:1234/path"))
+        self.loop.run_until_complete(connector.close())
 
     def test_proxy_auth_property(self) -> None:
         req = aiohttp.ClientRequest(
@@ -922,9 +935,7 @@ class TestProxy(unittest.TestCase):
                             tls_m.assert_called_with(
                                 mock.ANY,
                                 mock.ANY,
-                                self.loop.run_until_complete(
-                                    connector._make_or_get_ssl_context(True)
-                                ),
+                                _SSL_CONTEXT_VERIFIED,
                                 server_hostname="www.python.org",
                                 ssl_handshake_timeout=mock.ANY,
                             )
@@ -938,6 +949,7 @@ class TestProxy(unittest.TestCase):
                             self.loop.run_until_complete(proxy_req.close())
                             proxy_resp.close()
                             self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
 
     @mock.patch("aiohttp.connector.ClientRequest")
     @mock.patch(
@@ -1031,3 +1043,4 @@ class TestProxy(unittest.TestCase):
                             self.loop.run_until_complete(proxy_req.close())
                             proxy_resp.close()
                             self.loop.run_until_complete(req.close())
+                self.loop.run_until_complete(connector.close())
