@@ -1,15 +1,16 @@
-# type: ignore
 import asyncio
 import io
 import json
 import pathlib
 import socket
+import sys
 import zlib
-from typing import Any, Optional
+from typing import AsyncIterator, Awaitable, Callable, Dict, List, NoReturn, Optional
 from unittest import mock
 
 import pytest
 from multidict import CIMultiDictProxy, MultiDict
+from pytest_mock import MockerFixture
 from yarl import URL
 
 import aiohttp
@@ -22,9 +23,12 @@ from aiohttp import (
     multipart,
     web,
 )
+from aiohttp.abc import AbstractResolver, ResolveResult
 from aiohttp.hdrs import CONTENT_LENGTH, CONTENT_TYPE, TRANSFER_ENCODING
+from aiohttp.pytest_plugin import AiohttpClient, AiohttpServer
 from aiohttp.test_utils import make_mocked_coro
-from aiohttp.typedefs import Handler
+from aiohttp.typedefs import Handler, Middleware
+from aiohttp.web_protocol import RequestHandler
 
 try:
     import brotlicffi as brotli
@@ -34,27 +38,27 @@ except ImportError:
 try:
     import ssl
 except ImportError:
-    ssl = None
+    ssl = None  # type: ignore[assignment]
 
 
 @pytest.fixture
-def here():
+def here() -> pathlib.Path:
     return pathlib.Path(__file__).parent
 
 
 @pytest.fixture
-def fname(here: Any):
+def fname(here: pathlib.Path) -> pathlib.Path:
     return here / "conftest.py"
 
 
-def new_dummy_form():
+def new_dummy_form() -> FormData:
     form = FormData()
     form.add_field("name", b"123")
     return form
 
 
-async def test_simple_get(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_simple_get(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         body = await request.read()
         assert b"" == body
         return web.Response(body=b"OK")
@@ -71,8 +75,8 @@ async def test_simple_get(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_simple_get_with_text(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_simple_get_with_text(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         body = await request.read()
         assert b"" == body
         return web.Response(text="OK", headers={"content-type": "text/plain"})
@@ -90,50 +94,57 @@ async def test_simple_get_with_text(aiohttp_client: Any) -> None:
 
 
 async def test_handler_returns_not_response(
-    aiohttp_server: Any, aiohttp_client: Any
+    aiohttp_server: AiohttpServer, aiohttp_client: AiohttpClient
 ) -> None:
     asyncio.get_event_loop().set_debug(True)
     logger = mock.Mock()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> str:
         return "abc"
 
     app = web.Application()
-    app.router.add_get("/", handler)
+    app.router.add_get("/", handler)  # type: ignore[arg-type]
     server = await aiohttp_server(app, logger=logger)
     client = await aiohttp_client(server)
 
-    with pytest.raises(aiohttp.ServerDisconnectedError):
-        await client.get("/")
-
-    logger.exception.assert_called_with(
-        "Unhandled runtime exception", exc_info=mock.ANY
-    )
+    async with client.get("/") as resp:
+        assert resp.status == 500
 
 
-async def test_handler_returns_none(aiohttp_server: Any, aiohttp_client: Any) -> None:
+async def test_handler_returns_none(
+    aiohttp_server: AiohttpServer, aiohttp_client: AiohttpClient
+) -> None:
     asyncio.get_event_loop().set_debug(True)
     logger = mock.Mock()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> None:
         return None
 
     app = web.Application()
-    app.router.add_get("/", handler)
+    app.router.add_get("/", handler)  # type: ignore[arg-type]
     server = await aiohttp_server(app, logger=logger)
     client = await aiohttp_client(server)
 
-    with pytest.raises(aiohttp.ServerDisconnectedError):
-        await client.get("/")
-
-    # Actual error text is placed in exc_info
-    logger.exception.assert_called_with(
-        "Unhandled runtime exception", exc_info=mock.ANY
-    )
+    async with client.get("/") as resp:
+        assert resp.status == 500
 
 
-async def test_head_returns_empty_body(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_handler_returns_not_response_after_100expect(
+    aiohttp_server: AiohttpServer, aiohttp_client: AiohttpClient
+) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        raise Exception("foo")
+
+    app = web.Application()
+    app.router.add_get("/", handler)
+    client = await aiohttp_client(app)
+
+    async with client.get("/", expect100=True) as resp:
+        assert resp.status == 500
+
+
+async def test_head_returns_empty_body(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(body=b"test")
 
     app = web.Application()
@@ -150,8 +161,25 @@ async def test_head_returns_empty_body(aiohttp_client: Any) -> None:
     assert resp.headers["Content-Length"] == "4"
 
 
-async def test_response_before_complete(aiohttp_client: Any) -> None:
-    async def handler(request):
+@pytest.mark.parametrize("status", (201, 204, 404))
+async def test_default_content_type_no_body(
+    aiohttp_client: AiohttpClient, status: int
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
+        return web.Response(status=status)
+
+    app = web.Application()
+    app.router.add_get("/", handler)
+    client = await aiohttp_client(app)
+
+    async with client.get("/") as resp:
+        assert resp.status == status
+        assert await resp.read() == b""
+        assert "Content-Type" not in resp.headers
+
+
+async def test_response_before_complete(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(body=b"OK")
 
     app = web.Application()
@@ -168,8 +196,42 @@ async def test_response_before_complete(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_form(aiohttp_client: Any) -> None:
-    async def handler(request):
+@pytest.mark.skipif(sys.version_info < (3, 11), reason="Needs Task.cancelling()")
+async def test_cancel_shutdown(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
+        t = asyncio.create_task(request.protocol.shutdown())
+        # Ensure it's started waiting
+        await asyncio.sleep(0)
+
+        t.cancel()
+        # Cancellation should not be suppressed
+        with pytest.raises(asyncio.CancelledError):
+            await t
+
+        # Repeat for second waiter in shutdown()
+        with mock.patch.object(request.protocol, "_request_in_progress", False):
+            with mock.patch.object(request.protocol, "_current_request", None):
+                t = asyncio.create_task(request.protocol.shutdown())
+                await asyncio.sleep(0)
+
+                t.cancel()
+                with pytest.raises(asyncio.CancelledError):
+                    await t
+
+        return web.Response(body=b"OK")
+
+    app = web.Application()
+    app.router.add_get("/", handler)
+    client = await aiohttp_client(app)
+
+    async with client.get("/") as resp:
+        assert resp.status == 200
+        txt = await resp.text()
+        assert txt == "OK"
+
+
+async def test_post_form(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         assert {"a": "1", "b": "2", "c": ""} == data
         return web.Response(body=b"OK")
@@ -186,8 +248,8 @@ async def test_post_form(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_text(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_post_text(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         data = await request.text()
         assert "русский" == data
         data2 = await request.text()
@@ -206,10 +268,10 @@ async def test_post_text(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_json(aiohttp_client: Any) -> None:
+async def test_post_json(aiohttp_client: AiohttpClient) -> None:
     dct = {"key": "текст"}
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         data = await request.json()
         assert dct == data
         data2 = await request.json(loads=json.loads)
@@ -232,12 +294,12 @@ async def test_post_json(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_multipart(aiohttp_client: Any) -> None:
+async def test_multipart(aiohttp_client: AiohttpClient) -> None:
     with multipart.MultipartWriter() as writer:
         writer.append("test")
         writer.append_json({"passed": True})
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         reader = await request.multipart()
         assert isinstance(reader, multipart.MultipartReader)
 
@@ -249,8 +311,8 @@ async def test_multipart(aiohttp_client: Any) -> None:
         part = await reader.next()
         assert isinstance(part, multipart.BodyPartReader)
         assert part.headers["Content-Type"] == "application/json"
-        thing = await part.json()
-        assert thing == {"passed": True}
+        json_thing = await part.json()
+        assert json_thing == {"passed": True}
 
         resp = web.Response()
         resp.content_type = "application/json"
@@ -266,11 +328,11 @@ async def test_multipart(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_multipart_empty(aiohttp_client: Any) -> None:
+async def test_multipart_empty(aiohttp_client: AiohttpClient) -> None:
     with multipart.MultipartWriter() as writer:
         pass
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         reader = await request.multipart()
         assert isinstance(reader, multipart.MultipartReader)
         async for part in reader:
@@ -286,12 +348,17 @@ async def test_multipart_empty(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_multipart_content_transfer_encoding(aiohttp_client: Any) -> None:
+async def test_multipart_content_transfer_encoding(
+    aiohttp_client: AiohttpClient,
+) -> None:
     # For issue #1168
     with multipart.MultipartWriter() as writer:
-        writer.append(b"\x00" * 10, headers={"Content-Transfer-Encoding": "binary"})
+        writer.append(
+            b"\x00" * 10,
+            headers={"Content-Transfer-Encoding": "binary"},
+        )
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         reader = await request.multipart()
         assert isinstance(reader, multipart.MultipartReader)
 
@@ -315,8 +382,8 @@ async def test_multipart_content_transfer_encoding(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_render_redirect(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_render_redirect(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         raise web.HTTPMovedPermanently(location="/path")
 
     app = web.Application()
@@ -332,20 +399,21 @@ async def test_render_redirect(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_single_file(aiohttp_client: Any) -> None:
+async def test_post_single_file(aiohttp_client: AiohttpClient) -> None:
     here = pathlib.Path(__file__).parent
 
-    def check_file(fs):
+    def check_file(fs: aiohttp.web_request.FileField) -> None:
         fullname = here / fs.filename
         with fullname.open("rb") as f:
             test_data = f.read()
             data = fs.file.read()
             assert test_data == data
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         assert ["data.unknown_mime_type"] == list(data.keys())
         for fs in data.values():
+            assert isinstance(fs, aiohttp.web_request.FileField)
             check_file(fs)
             fs.file.close()
         resp = web.Response(body=b"OK")
@@ -364,12 +432,13 @@ async def test_post_single_file(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_files_upload_with_same_key(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_files_upload_with_same_key(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         files = data.getall("file")
         file_names = set()
         for _file in files:
+            assert isinstance(_file, aiohttp.web_request.FileField)
             assert not _file.file.closed
             if _file.filename == "test1.jpeg":
                 assert _file.file.read() == b"binary data 1"
@@ -399,20 +468,21 @@ async def test_files_upload_with_same_key(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_files(aiohttp_client: Any) -> None:
+async def test_post_files(aiohttp_client: AiohttpClient) -> None:
     here = pathlib.Path(__file__).parent
 
-    def check_file(fs):
+    def check_file(fs: aiohttp.web_request.FileField) -> None:
         fullname = here / fs.filename
         with fullname.open("rb") as f:
             test_data = f.read()
             data = fs.file.read()
             assert test_data == data
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         assert ["data.unknown_mime_type", "conftest.py"] == list(data.keys())
         for fs in data.values():
+            assert isinstance(fs, aiohttp.web_request.FileField)
             check_file(fs)
             fs.file.close()
         resp = web.Response(body=b"OK")
@@ -430,8 +500,8 @@ async def test_post_files(aiohttp_client: Any) -> None:
             resp.release()
 
 
-async def test_release_post_data(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_release_post_data(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         await request.release()
         chunk = await request.content.readany()
         assert chunk == b""
@@ -447,8 +517,8 @@ async def test_release_post_data(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_form_with_duplicate_keys(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_post_form_with_duplicate_keys(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         lst = list(data.items())
         assert [("a", "1"), ("a", "2")] == lst
@@ -469,7 +539,7 @@ def test_repr_for_application() -> None:
     assert f"<Application 0x{id(app):x}>" == repr(app)
 
 
-async def test_expect_default_handler_unknown(aiohttp_client: Any) -> None:
+async def test_expect_default_handler_unknown(aiohttp_client: AiohttpClient) -> None:
     # Test default Expect handler for unknown Expect value.
 
     # A server that does not understand or is unable to comply with any of
@@ -480,12 +550,8 @@ async def test_expect_default_handler_unknown(aiohttp_client: Any) -> None:
     # status.
 
     # http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.20
-    async def handler(request):
-        await request.post()
-        pytest.xfail(
-            "Handler should not proceed to this point in case of "
-            "unknown Expect header"
-        )
+    async def handler(request: web.Request) -> web.Response:
+        assert False
 
     app = web.Application()
     app.router.add_post("/", handler)
@@ -497,8 +563,8 @@ async def test_expect_default_handler_unknown(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_100_continue(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_100_continue(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         assert b"123" == data["name"]
         return web.Response()
@@ -516,19 +582,19 @@ async def test_100_continue(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_100_continue_custom(aiohttp_client: Any) -> None:
+async def test_100_continue_custom(aiohttp_client: AiohttpClient) -> None:
     expect_received = False
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         assert b"123" == data["name"]
         return web.Response()
 
-    async def expect_handler(request):
+    async def expect_handler(request: web.Request) -> None:
         nonlocal expect_received
         expect_received = True
-        if request.version == HttpVersion11:
-            await request.writer.write(b"HTTP/1.1 100 Continue\r\n\r\n")
+        assert request.version == HttpVersion11
+        await request.writer.write(b"HTTP/1.1 100 Continue\r\n\r\n")
 
     app = web.Application()
     app.router.add_post("/", handler, expect_handler=expect_handler)
@@ -541,18 +607,18 @@ async def test_100_continue_custom(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_100_continue_custom_response(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_100_continue_custom_response(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         data = await request.post()
         assert b"123", data["name"]
         return web.Response()
 
-    async def expect_handler(request):
-        if request.version == HttpVersion11:
-            if auth_err:
-                raise web.HTTPForbidden()
+    async def expect_handler(request: web.Request) -> None:
+        assert request.version == HttpVersion11
+        if auth_err:
+            raise web.HTTPForbidden()
 
-            await request.writer.write(b"HTTP/1.1 100 Continue\r\n\r\n")
+        await request.writer.write(b"HTTP/1.1 100 Continue\r\n\r\n")
 
     app = web.Application()
     app.router.add_post("/", handler, expect_handler=expect_handler)
@@ -569,17 +635,16 @@ async def test_100_continue_custom_response(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_expect_handler_custom_response(aiohttp_client: Any) -> None:
+async def test_expect_handler_custom_response(aiohttp_client: AiohttpClient) -> None:
     cache = {"foo": "bar"}
 
     async def handler(request: web.Request) -> web.Response:
         return web.Response(text="handler")
 
     async def expect_handler(request: web.Request) -> Optional[web.Response]:
-        k = request.headers.get("X-Key")
+        k = request.headers["X-Key"]
         cached_value = cache.get(k)
-        if cached_value:
-            return web.Response(text=cached_value)
+        return web.Response(text=cached_value) if cached_value else None
 
     app = web.Application()
     # expect_handler is only typed on add_route().
@@ -595,7 +660,7 @@ async def test_expect_handler_custom_response(aiohttp_client: Any) -> None:
         assert await resp.text() == "handler"
 
 
-async def test_100_continue_for_not_found(aiohttp_client: Any) -> None:
+async def test_100_continue_for_not_found(aiohttp_client: AiohttpClient) -> None:
     app = web.Application()
     client = await aiohttp_client(app)
 
@@ -605,9 +670,9 @@ async def test_100_continue_for_not_found(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_100_continue_for_not_allowed(aiohttp_client: Any) -> None:
-    async def handler(request):
-        return web.Response()
+async def test_100_continue_for_not_allowed(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     app = web.Application()
     app.router.add_post("/", handler)
@@ -619,8 +684,8 @@ async def test_100_continue_for_not_allowed(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_http11_keep_alive_default(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_http11_keep_alive_default(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -636,24 +701,24 @@ async def test_http11_keep_alive_default(aiohttp_client: Any) -> None:
 
 
 @pytest.mark.xfail
-async def test_http10_keep_alive_default(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_http10_keep_alive_default(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
     app.router.add_get("/", handler)
     client = await aiohttp_client(app, version=HttpVersion10)
 
-    resp = await client.get("/")
-    assert 200 == resp.status
-    assert resp.version == HttpVersion10
-    assert resp.headers["Connection"] == "keep-alive"
-
-    resp.release()
+    async with client.get("/") as resp:
+        assert 200 == resp.status
+        assert resp.version == HttpVersion10
+        assert resp.headers["Connection"] == "keep-alive"
 
 
-async def test_http10_keep_alive_with_headers_close(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_http10_keep_alive_with_headers_close(
+    aiohttp_client: AiohttpClient,
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
         await request.read()
         return web.Response(body=b"OK")
 
@@ -670,8 +735,8 @@ async def test_http10_keep_alive_with_headers_close(aiohttp_client: Any) -> None
     resp.release()
 
 
-async def test_http10_keep_alive_with_headers(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_http10_keep_alive_with_headers(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         await request.read()
         return web.Response(body=b"OK")
 
@@ -688,16 +753,18 @@ async def test_http10_keep_alive_with_headers(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_upload_file(aiohttp_client: Any) -> None:
+async def test_upload_file(aiohttp_client: AiohttpClient) -> None:
     here = pathlib.Path(__file__).parent
     fname = here / "aiohttp.png"
     with fname.open("rb") as f:
         data = f.read()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         form = await request.post()
-        raw_data = form["file"].file.read()
-        form["file"].file.close()
+        form_file = form["file"]
+        assert isinstance(form_file, aiohttp.web_request.FileField)
+        raw_data = form_file.file.read()
+        form_file.file.close()
         assert data == raw_data
         return web.Response()
 
@@ -711,16 +778,18 @@ async def test_upload_file(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_upload_file_object(aiohttp_client: Any) -> None:
+async def test_upload_file_object(aiohttp_client: AiohttpClient) -> None:
     here = pathlib.Path(__file__).parent
     fname = here / "aiohttp.png"
     with fname.open("rb") as f:
         data = f.read()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         form = await request.post()
-        raw_data = form["file"].file.read()
-        form["file"].file.close()
+        form_file = form["file"]
+        assert isinstance(form_file, aiohttp.web_request.FileField)
+        raw_data = form_file.file.read()
+        form_file.file.close()
         assert data == raw_data
         return web.Response()
 
@@ -739,9 +808,9 @@ async def test_upload_file_object(aiohttp_client: Any) -> None:
     "method", ["get", "post", "options", "post", "put", "patch", "delete"]
 )
 async def test_empty_content_for_query_without_body(
-    method: Any, aiohttp_client: Any
+    method: str, aiohttp_client: AiohttpClient
 ) -> None:
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         assert not request.body_exists
         assert not request.can_read_body
         return web.Response()
@@ -754,8 +823,8 @@ async def test_empty_content_for_query_without_body(
     assert 200 == resp.status
 
 
-async def test_empty_content_for_query_with_body(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_empty_content_for_query_with_body(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert request.body_exists
         assert request.can_read_body
         body = await request.read()
@@ -771,8 +840,8 @@ async def test_empty_content_for_query_with_body(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_get_with_empty_arg(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_get_with_empty_arg(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert "arg" in request.query
         assert "" == request.query["arg"]
         return web.Response()
@@ -787,9 +856,9 @@ async def test_get_with_empty_arg(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_large_header(aiohttp_client: Any) -> None:
-    async def handler(request):
-        return web.Response()
+async def test_large_header(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     app = web.Application()
     app.router.add_get("/", handler)
@@ -802,8 +871,10 @@ async def test_large_header(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_large_header_allowed(aiohttp_client: Any, aiohttp_server: Any) -> None:
-    async def handler(request):
+async def test_large_header_allowed(
+    aiohttp_client: AiohttpClient, aiohttp_server: AiohttpServer
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -818,8 +889,8 @@ async def test_large_header_allowed(aiohttp_client: Any, aiohttp_server: Any) ->
     resp.release()
 
 
-async def test_get_with_empty_arg_with_equal(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_get_with_empty_arg_with_equal(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert "arg" in request.query
         assert "" == request.query["arg"]
         return web.Response()
@@ -834,20 +905,22 @@ async def test_get_with_empty_arg_with_equal(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_response_with_async_gen(aiohttp_client: Any, fname: Any) -> None:
+async def test_response_with_async_gen(
+    aiohttp_client: AiohttpClient, fname: pathlib.Path
+) -> None:
     with fname.open("rb") as f:
         data = f.read()
 
     data_size = len(data)
 
-    async def stream(f_name):
+    async def stream(f_name: pathlib.Path) -> AsyncIterator[bytes]:
         with f_name.open("rb") as f:
             data = f.read(100)
             while data:
                 yield data
                 data = f.read(100)
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         headers = {"Content-Length": str(data_size)}
         return web.Response(body=stream(fname), headers=headers)
 
@@ -865,21 +938,21 @@ async def test_response_with_async_gen(aiohttp_client: Any, fname: Any) -> None:
 
 
 async def test_response_with_async_gen_no_params(
-    aiohttp_client: Any, fname: Any
+    aiohttp_client: AiohttpClient, fname: pathlib.Path
 ) -> None:
     with fname.open("rb") as f:
         data = f.read()
 
     data_size = len(data)
 
-    async def stream():
+    async def stream() -> AsyncIterator[bytes]:
         with fname.open("rb") as f:
             data = f.read(100)
             while data:
                 yield data
                 data = f.read(100)
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         headers = {"Content-Length": str(data_size)}
         return web.Response(body=stream(), headers=headers)
 
@@ -896,13 +969,15 @@ async def test_response_with_async_gen_no_params(
     resp.release()
 
 
-async def test_response_with_file(aiohttp_client: Any, fname: Any) -> None:
+async def test_response_with_file(
+    aiohttp_client: AiohttpClient, fname: pathlib.Path
+) -> None:
     outer_file_descriptor = None
 
     with fname.open("rb") as f:
         data = f.read()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         nonlocal outer_file_descriptor
         outer_file_descriptor = fname.open("rb")
         return web.Response(body=outer_file_descriptor)
@@ -926,16 +1001,19 @@ async def test_response_with_file(aiohttp_client: Any, fname: Any) -> None:
 
     resp.release()
 
+    assert outer_file_descriptor is not None
     outer_file_descriptor.close()
 
 
-async def test_response_with_file_ctype(aiohttp_client: Any, fname: Any) -> None:
+async def test_response_with_file_ctype(
+    aiohttp_client: AiohttpClient, fname: pathlib.Path
+) -> None:
     outer_file_descriptor = None
 
     with fname.open("rb") as f:
         data = f.read()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         nonlocal outer_file_descriptor
         outer_file_descriptor = fname.open("rb")
 
@@ -958,16 +1036,19 @@ async def test_response_with_file_ctype(aiohttp_client: Any, fname: Any) -> None
 
     resp.release()
 
+    assert outer_file_descriptor is not None
     outer_file_descriptor.close()
 
 
-async def test_response_with_payload_disp(aiohttp_client: Any, fname: Any) -> None:
+async def test_response_with_payload_disp(
+    aiohttp_client: AiohttpClient, fname: pathlib.Path
+) -> None:
     outer_file_descriptor = None
 
     with fname.open("rb") as f:
         data = f.read()
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         nonlocal outer_file_descriptor
         outer_file_descriptor = fname.open("rb")
         pl = aiohttp.get_payload(outer_file_descriptor)
@@ -988,11 +1069,14 @@ async def test_response_with_payload_disp(aiohttp_client: Any, fname: Any) -> No
 
     resp.release()
 
+    assert outer_file_descriptor is not None
     outer_file_descriptor.close()
 
 
-async def test_response_with_payload_stringio(aiohttp_client: Any, fname: Any) -> None:
-    async def handler(request):
+async def test_response_with_payload_stringio(
+    aiohttp_client: AiohttpClient, fname: pathlib.Path
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(body=io.StringIO("test"))
 
     app = web.Application()
@@ -1018,9 +1102,9 @@ async def test_response_with_payload_stringio(aiohttp_client: Any, fname: Any) -
     ],
 )
 async def test_response_with_precompressed_body(
-    aiohttp_client: Any, compressor: Any, encoding: Any
+    aiohttp_client: AiohttpClient, compressor: "zlib._Compress", encoding: str
 ) -> None:
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         headers = {"Content-Encoding": encoding}
         data = compressor.compress(b"mydata") + compressor.flush()
         return web.Response(body=data, headers=headers)
@@ -1038,8 +1122,10 @@ async def test_response_with_precompressed_body(
     resp.release()
 
 
-async def test_response_with_precompressed_body_brotli(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_response_with_precompressed_body_brotli(
+    aiohttp_client: AiohttpClient,
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
         headers = {"Content-Encoding": "br"}
         return web.Response(body=brotli.compress(b"mydata"), headers=headers)
 
@@ -1056,8 +1142,8 @@ async def test_response_with_precompressed_body_brotli(aiohttp_client: Any) -> N
     resp.release()
 
 
-async def test_bad_request_payload(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_bad_request_payload(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert request.method == "POST"
 
         with pytest.raises(aiohttp.web.RequestPayloadError):
@@ -1075,8 +1161,8 @@ async def test_bad_request_payload(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_stream_response_multiple_chunks(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_stream_response_multiple_chunks(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.StreamResponse:
         resp = web.StreamResponse()
         resp.enable_chunked_encoding()
         await resp.prepare(request)
@@ -1097,7 +1183,7 @@ async def test_stream_response_multiple_chunks(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_start_without_routes(aiohttp_client: Any) -> None:
+async def test_start_without_routes(aiohttp_client: AiohttpClient) -> None:
     app = web.Application()
     client = await aiohttp_client(app)
 
@@ -1107,8 +1193,8 @@ async def test_start_without_routes(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_requests_count(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_requests_count(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -1132,11 +1218,11 @@ async def test_requests_count(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_redirect_url(aiohttp_client: Any) -> None:
-    async def redirector(request):
+async def test_redirect_url(aiohttp_client: AiohttpClient) -> None:
+    async def redirector(request: web.Request) -> NoReturn:
         raise web.HTTPFound(location=URL("/redirected"))
 
-    async def redirected(request):
+    async def redirected(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -1150,8 +1236,8 @@ async def test_redirect_url(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_simple_subapp(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_simple_subapp(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(text="OK")
 
     app = web.Application()
@@ -1168,11 +1254,11 @@ async def test_simple_subapp(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_reverse_url(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_subapp_reverse_url(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         raise web.HTTPMovedPermanently(location=subapp.router["name"].url_for())
 
-    async def handler2(request):
+    async def handler2(request: web.Request) -> web.Response:
         return web.Response(text="OK")
 
     app = web.Application()
@@ -1191,13 +1277,13 @@ async def test_subapp_reverse_url(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_reverse_variable_url(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_subapp_reverse_variable_url(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         raise web.HTTPMovedPermanently(
             location=subapp.router["name"].url_for(part="final")
         )
 
-    async def handler2(request):
+    async def handler2(request: web.Request) -> web.Response:
         return web.Response(text="OK")
 
     app = web.Application()
@@ -1216,10 +1302,10 @@ async def test_subapp_reverse_variable_url(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_reverse_static_url(aiohttp_client: Any) -> None:
+async def test_subapp_reverse_static_url(aiohttp_client: AiohttpClient) -> None:
     fname = "aiohttp.png"
 
-    async def handler(request):
+    async def handler(request: web.Request) -> NoReturn:
         raise web.HTTPMovedPermanently(
             location=subapp.router["name"].url_for(filename=fname)
         )
@@ -1243,8 +1329,8 @@ async def test_subapp_reverse_static_url(aiohttp_client: Any) -> None:
         assert body == f.read()
 
 
-async def test_subapp_app(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_subapp_app(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert request.app is subapp
         return web.Response(text="OK")
 
@@ -1262,9 +1348,9 @@ async def test_subapp_app(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_not_found(aiohttp_client: Any) -> None:
-    async def handler(request):
-        return web.Response(text="OK")
+async def test_subapp_not_found(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     app = web.Application()
     subapp = web.Application()
@@ -1278,9 +1364,9 @@ async def test_subapp_not_found(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_not_found2(aiohttp_client: Any) -> None:
-    async def handler(request):
-        return web.Response(text="OK")
+async def test_subapp_not_found2(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     app = web.Application()
     subapp = web.Application()
@@ -1294,9 +1380,9 @@ async def test_subapp_not_found2(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_not_allowed(aiohttp_client: Any) -> None:
-    async def handler(request):
-        return web.Response(text="OK")
+async def test_subapp_not_allowed(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
+        assert False
 
     app = web.Application()
     subapp = web.Application()
@@ -1311,10 +1397,10 @@ async def test_subapp_not_allowed(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_cannot_add_app_in_handler(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_subapp_cannot_add_app_in_handler(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         request.match_info.add_app(app)
-        return web.Response(text="OK")
+        assert False
 
     app = web.Application()
     subapp = web.Application()
@@ -1328,10 +1414,10 @@ async def test_subapp_cannot_add_app_in_handler(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_old_style_subapp_middlewares(aiohttp_client: Any) -> None:
+async def test_old_style_subapp_middlewares(aiohttp_client: AiohttpClient) -> None:
     order = []
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(text="OK")
 
     with pytest.deprecated_call(
@@ -1341,7 +1427,9 @@ async def test_old_style_subapp_middlewares(aiohttp_client: Any) -> None:
     ):
 
         @web.middleware
-        async def middleware(request, handler: Handler):
+        async def middleware(
+            request: web.Request, handler: Handler
+        ) -> web.StreamResponse:
             order.append((1, request.app[name]))
             resp = await handler(request)
             assert 200 == resp.status
@@ -1375,14 +1463,18 @@ async def test_old_style_subapp_middlewares(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_on_response_prepare(aiohttp_client: Any) -> None:
+async def test_subapp_on_response_prepare(aiohttp_client: AiohttpClient) -> None:
     order = []
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(text="OK")
 
-    def make_signal(app):
-        async def on_response(request, response):
+    def make_signal(
+        app: web.Application,
+    ) -> Callable[[web.Request, web.StreamResponse], Awaitable[None]]:
+        async def on_response(
+            request: web.Request, response: web.StreamResponse
+        ) -> None:
             order.append(app)
 
         return on_response
@@ -1405,10 +1497,10 @@ async def test_subapp_on_response_prepare(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_subapp_on_startup(aiohttp_server: Any) -> None:
+async def test_subapp_on_startup(aiohttp_server: AiohttpServer) -> None:
     order = []
 
-    async def on_signal(app):
+    async def on_signal(app: web.Application) -> None:
         order.append(app)
 
     app = web.Application()
@@ -1425,10 +1517,10 @@ async def test_subapp_on_startup(aiohttp_server: Any) -> None:
     assert [app, subapp1, subapp2] == order
 
 
-async def test_subapp_on_shutdown(aiohttp_server: Any) -> None:
+async def test_subapp_on_shutdown(aiohttp_server: AiohttpServer) -> None:
     order = []
 
-    async def on_signal(app):
+    async def on_signal(app: web.Application) -> None:
         order.append(app)
 
     app = web.Application()
@@ -1446,10 +1538,10 @@ async def test_subapp_on_shutdown(aiohttp_server: Any) -> None:
     assert [app, subapp1, subapp2] == order
 
 
-async def test_subapp_on_cleanup(aiohttp_server: Any) -> None:
+async def test_subapp_on_cleanup(aiohttp_server: AiohttpServer) -> None:
     order = []
 
-    async def on_signal(app):
+    async def on_signal(app: web.Application) -> None:
         order.append(app)
 
     app = web.Application()
@@ -1481,19 +1573,21 @@ async def test_subapp_on_cleanup(aiohttp_server: Any) -> None:
     ],
 )
 async def test_subapp_middleware_context(
-    aiohttp_client: Any, route: Any, expected: Any, middlewares: Any
-):
+    aiohttp_client: AiohttpClient, route: str, expected: List[str], middlewares: str
+) -> None:
     values = []
 
-    def show_app_context(appname):
-        async def middleware(request, handler: Handler):
+    def show_app_context(appname: str) -> Middleware:
+        async def middleware(
+            request: web.Request, handler: Handler
+        ) -> web.StreamResponse:
             values.append(f"{appname}: {request.app[my_value]}")
             return await handler(request)
 
         return middleware
 
-    def make_handler(appname):
-        async def handler(request):
+    def make_handler(appname: str) -> Handler:
+        async def handler(request: web.Request) -> web.Response:
             values.append(f"{appname}: {request.app[my_value]}")
             return web.Response(text="Ok")
 
@@ -1522,8 +1616,8 @@ async def test_subapp_middleware_context(
     resp.release()
 
 
-async def test_custom_date_header(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_custom_date_header(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(headers={"Date": "Sun, 30 Oct 2016 03:13:52 GMT"})
 
     app = web.Application()
@@ -1537,8 +1631,8 @@ async def test_custom_date_header(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_response_prepared_with_clone(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_response_prepared_with_clone(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.StreamResponse:
         cloned = request.clone()
         resp = web.StreamResponse()
         await resp.prepare(cloned)
@@ -1554,10 +1648,10 @@ async def test_response_prepared_with_clone(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_app_max_client_size(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_app_max_client_size(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         await request.post()
-        return web.Response(body=b"ok")
+        assert False
 
     max_size = 1024**2
     app = web.Application()
@@ -1568,9 +1662,7 @@ async def test_app_max_client_size(aiohttp_client: Any) -> None:
         resp = await client.post("/", data=data)
     assert 413 == resp.status
     resp_text = await resp.text()
-    assert (
-        "Maximum request body size 1048576 exceeded, " "actual body size" in resp_text
-    )
+    assert "Maximum request body size 1048576 exceeded, actual body size" in resp_text
     # Maximum request body size X exceeded, actual body size X
     body_size = int(resp_text.split()[-1])
     assert body_size >= max_size
@@ -1578,8 +1670,8 @@ async def test_app_max_client_size(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_app_max_client_size_adjusted(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_app_max_client_size_adjusted(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         await request.post()
         return web.Response(body=b"ok")
 
@@ -1602,9 +1694,7 @@ async def test_app_max_client_size_adjusted(aiohttp_client: Any) -> None:
         resp = await client.post("/", data=too_large_data)
     assert 413 == resp.status
     resp_text = await resp.text()
-    assert (
-        "Maximum request body size 2097152 exceeded, " "actual body size" in resp_text
-    )
+    assert "Maximum request body size 2097152 exceeded, actual body size" in resp_text
     # Maximum request body size X exceeded, actual body size X
     body_size = int(resp_text.split()[-1])
     assert body_size >= custom_max_size
@@ -1612,14 +1702,13 @@ async def test_app_max_client_size_adjusted(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_app_max_client_size_none(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_app_max_client_size_none(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         await request.post()
         return web.Response(body=b"ok")
 
     default_max_size = 1024**2
-    custom_max_size = None
-    app = web.Application(client_max_size=custom_max_size)
+    app = web.Application(client_max_size=0)
     app.router.add_post("/", handler)
     client = await aiohttp_client(app)
 
@@ -1640,10 +1729,10 @@ async def test_app_max_client_size_none(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_post_max_client_size(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_post_max_client_size(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         await request.post()
-        return web.Response()
+        assert False
 
     app = web.Application(client_max_size=10)
     app.router.add_post("/", handler)
@@ -1659,15 +1748,17 @@ async def test_post_max_client_size(aiohttp_client: Any) -> None:
             "Maximum request body size 10 exceeded, "
             "actual body size 1024" in resp_text
         )
-        data["file"].close()
+        data_file = data["file"]
+        assert isinstance(data_file, io.BytesIO)
+        data_file.close()
 
         resp.release()
 
 
-async def test_post_max_client_size_for_file(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_post_max_client_size_for_file(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         await request.post()
-        return web.Response()
+        assert False
 
     app = web.Application(client_max_size=2)
     app.router.add_post("/", handler)
@@ -1682,8 +1773,8 @@ async def test_post_max_client_size_for_file(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_response_with_bodypart(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_response_with_bodypart(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         reader = await request.multipart()
         part = await reader.next()
         return web.Response(body=part)
@@ -1706,8 +1797,10 @@ async def test_response_with_bodypart(aiohttp_client: Any) -> None:
         resp.release()
 
 
-async def test_response_with_bodypart_named(aiohttp_client: Any, tmp_path: Any) -> None:
-    async def handler(request):
+async def test_response_with_bodypart_named(
+    aiohttp_client: AiohttpClient, tmp_path: pathlib.Path
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
         reader = await request.multipart()
         part = await reader.next()
         return web.Response(body=part)
@@ -1732,8 +1825,10 @@ async def test_response_with_bodypart_named(aiohttp_client: Any, tmp_path: Any) 
     resp.release()
 
 
-async def test_response_with_bodypart_invalid_name(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_response_with_bodypart_invalid_name(
+    aiohttp_client: AiohttpClient,
+) -> None:
+    async def handler(request: web.Request) -> web.Response:
         reader = await request.multipart()
         part = await reader.next()
         return web.Response(body=part)
@@ -1755,8 +1850,8 @@ async def test_response_with_bodypart_invalid_name(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_request_clone(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_request_clone(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         r2 = request.clone(method="POST")
         assert r2.method == "POST"
         assert r2.match_info is request.match_info
@@ -1771,8 +1866,8 @@ async def test_request_clone(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_await(aiohttp_server: Any) -> None:
-    async def handler(request):
+async def test_await(aiohttp_server: AiohttpServer) -> None:
+    async def handler(request: web.Request) -> web.StreamResponse:
         resp = web.StreamResponse(headers={"content-length": str(4)})
         await resp.prepare(request)
         with pytest.deprecated_call(
@@ -1798,8 +1893,8 @@ async def test_await(aiohttp_server: Any) -> None:
         assert resp.connection is None
 
 
-async def test_response_context_manager(aiohttp_server: Any) -> None:
-    async def handler(request):
+async def test_response_context_manager(aiohttp_server: AiohttpServer) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -1814,8 +1909,8 @@ async def test_response_context_manager(aiohttp_server: Any) -> None:
     await session.close()
 
 
-async def test_response_context_manager_error(aiohttp_server: Any) -> None:
-    async def handler(request):
+async def test_response_context_manager_error(aiohttp_server: AiohttpServer) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(text="some text")
 
     app = web.Application()
@@ -1831,13 +1926,14 @@ async def test_response_context_manager_error(aiohttp_server: Any) -> None:
             await resp.read()
     assert resp.closed
 
+    assert session._connector is not None
     assert len(session._connector._conns) == 1
 
     await session.close()
 
 
-async def aiohttp_client_api_context_manager(aiohttp_server: Any):
-    async def handler(request):
+async def test_client_api_context_manager(aiohttp_server: AiohttpServer) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -1852,9 +1948,9 @@ async def aiohttp_client_api_context_manager(aiohttp_server: Any):
 
 
 async def test_context_manager_close_on_release(
-    aiohttp_server: Any, mocker: Any
+    aiohttp_server: AiohttpServer, mocker: MockerFixture
 ) -> None:
-    async def handler(request):
+    async def handler(request: web.Request) -> web.StreamResponse:
         resp = web.StreamResponse()
         await resp.prepare(request)
         with pytest.deprecated_call(
@@ -1870,6 +1966,7 @@ async def test_context_manager_close_on_release(
 
     async with aiohttp.ClientSession() as session:
         resp = await session.get(server.make_url("/"))
+        assert resp.connection is not None
         proto = resp.connection._protocol
         mocker.spy(proto, "close")
         async with resp:
@@ -1881,10 +1978,10 @@ async def test_context_manager_close_on_release(
         resp.release()  # Trigger handler completion
 
 
-async def test_iter_any(aiohttp_server: Any) -> None:
+async def test_iter_any(aiohttp_server: AiohttpServer) -> None:
     data = b"0123456789" * 1024
 
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         buf = []
         async for raw in request.content.iter_any():
             buf.append(raw)
@@ -1900,7 +1997,7 @@ async def test_iter_any(aiohttp_server: Any) -> None:
             assert resp.status == 200
 
 
-async def test_request_tracing(aiohttp_server: Any) -> None:
+async def test_request_tracing(aiohttp_server: AiohttpServer) -> None:
     on_request_start = mock.Mock(side_effect=make_mocked_coro(mock.Mock()))
     on_request_end = mock.Mock(side_effect=make_mocked_coro(mock.Mock()))
     on_dns_resolvehost_start = mock.Mock(side_effect=make_mocked_coro(mock.Mock()))
@@ -1909,10 +2006,10 @@ async def test_request_tracing(aiohttp_server: Any) -> None:
     on_connection_create_start = mock.Mock(side_effect=make_mocked_coro(mock.Mock()))
     on_connection_create_end = mock.Mock(side_effect=make_mocked_coro(mock.Mock()))
 
-    async def redirector(request):
+    async def redirector(request: web.Request) -> NoReturn:
         raise web.HTTPFound(location=URL("/redirected"))
 
-    async def redirected(request):
+    async def redirected(request: web.Request) -> web.Response:
         return web.Response()
 
     trace_config = TraceConfig()
@@ -1930,29 +2027,35 @@ async def test_request_tracing(aiohttp_server: Any) -> None:
     app.router.add_get("/redirected", redirected)
     server = await aiohttp_server(app)
 
-    class FakeResolver:
+    class FakeResolver(AbstractResolver):
         _LOCAL_HOST = {0: "127.0.0.1", socket.AF_INET: "127.0.0.1"}
 
-        def __init__(self, fakes):
+        def __init__(self, fakes: Dict[str, int]):
             # fakes -- dns -> port dict
             self._fakes = fakes
             self._resolver = aiohttp.DefaultResolver()
 
-        async def resolve(self, host, port=0, family=socket.AF_INET):
+        async def close(self) -> None:
+            assert False
+
+        async def resolve(
+            self,
+            host: str,
+            port: int = 0,
+            family: socket.AddressFamily = socket.AF_INET,
+        ) -> List[ResolveResult]:
             fake_port = self._fakes.get(host)
-            if fake_port is not None:
-                return [
-                    {
-                        "hostname": host,
-                        "host": self._LOCAL_HOST[family],
-                        "port": fake_port,
-                        "family": socket.AF_INET,
-                        "proto": 0,
-                        "flags": socket.AI_NUMERICHOST,
-                    }
-                ]
-            else:
-                return await self._resolver.resolve(host, port, family)
+            assert fake_port is not None
+            return [
+                {
+                    "hostname": host,
+                    "host": self._LOCAL_HOST[family],
+                    "port": fake_port,
+                    "family": socket.AF_INET,
+                    "proto": 0,
+                    "flags": socket.AI_NUMERICHOST,
+                }
+            ]
 
     resolver = FakeResolver({"example.com": server.port})
     connector = aiohttp.TCPConnector(resolver=resolver)
@@ -1972,8 +2075,8 @@ async def test_request_tracing(aiohttp_server: Any) -> None:
     await client.close()
 
 
-async def test_raise_http_exception(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_raise_http_exception(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         raise web.HTTPForbidden()
 
     app = web.Application()
@@ -1985,8 +2088,8 @@ async def test_raise_http_exception(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_request_path(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_request_path(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert request.path_qs == "/path%20to?a=1"
         assert request.path == "/path to"
         assert request.raw_path == "/path%20to?a=1"
@@ -2003,8 +2106,8 @@ async def test_request_path(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_app_add_routes(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_app_add_routes(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         return web.Response()
 
     app = web.Application()
@@ -2016,8 +2119,8 @@ async def test_app_add_routes(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_request_headers_type(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_request_headers_type(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         assert isinstance(request.headers, CIMultiDictProxy)
         return web.Response()
 
@@ -2030,8 +2133,8 @@ async def test_request_headers_type(aiohttp_client: Any) -> None:
     resp.release()
 
 
-async def test_signal_on_error_handler(aiohttp_client: Any) -> None:
-    async def on_prepare(request, response):
+async def test_signal_on_error_handler(aiohttp_client: AiohttpClient) -> None:
+    async def on_prepare(request: web.Request, response: web.StreamResponse) -> None:
         response.headers["X-Custom"] = "val"
 
     app = web.Application()
@@ -2048,7 +2151,9 @@ async def test_signal_on_error_handler(aiohttp_client: Any) -> None:
     "HttpRequestParserC" not in dir(aiohttp.http_parser),
     reason="C based HTTP parser not available",
 )
-async def test_bad_method_for_c_http_parser_not_hangs(aiohttp_client: Any) -> None:
+async def test_bad_method_for_c_http_parser_not_hangs(
+    aiohttp_client: AiohttpClient,
+) -> None:
     app = web.Application()
     timeout = aiohttp.ClientTimeout(sock_read=0.2)
     client = await aiohttp_client(app, timeout=timeout)
@@ -2056,8 +2161,8 @@ async def test_bad_method_for_c_http_parser_not_hangs(aiohttp_client: Any) -> No
     assert 400 == resp.status
 
 
-async def test_read_bufsize(aiohttp_client: Any) -> None:
-    async def handler(request):
+async def test_read_bufsize(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.Response:
         ret = request.content.get_read_buffer_limits()
         data = await request.text()  # read posted data
         return web.Response(text=f"{data} {ret!r}")
@@ -2076,11 +2181,11 @@ async def test_read_bufsize(aiohttp_client: Any) -> None:
     "auto_decompress,len_of", [(True, "uncompressed"), (False, "compressed")]
 )
 async def test_auto_decompress(
-    aiohttp_client: Any,
+    aiohttp_client: AiohttpClient,
     auto_decompress: bool,
     len_of: str,
 ) -> None:
-    async def handler(request):
+    async def handler(request: web.Request) -> web.Response:
         data = await request.read()
         return web.Response(text=str(len(data)))
 
@@ -2104,9 +2209,9 @@ async def test_auto_decompress(
     [101, 204],
 )
 async def test_response_101_204_no_content_length_http11(
-    status: Any, aiohttp_client: Any
+    status: int, aiohttp_client: AiohttpClient
 ) -> None:
-    async def handler(_):
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(status=status)
 
     app = web.Application()
@@ -2118,8 +2223,8 @@ async def test_response_101_204_no_content_length_http11(
     resp.release()
 
 
-async def test_stream_response_headers_204(aiohttp_client: Any):
-    async def handler(_):
+async def test_stream_response_headers_204(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> web.StreamResponse:
         return web.StreamResponse(status=204)
 
     app = web.Application()
@@ -2131,8 +2236,8 @@ async def test_stream_response_headers_204(aiohttp_client: Any):
     resp.release()
 
 
-async def test_httpfound_cookies_302(aiohttp_client: Any) -> None:
-    async def handler(_):
+async def test_httpfound_cookies_302(aiohttp_client: AiohttpClient) -> None:
+    async def handler(request: web.Request) -> NoReturn:
         resp = web.HTTPFound("/")
         resp.set_cookie("my-cookie", "cookie-value")
         raise resp
@@ -2149,11 +2254,11 @@ async def test_httpfound_cookies_302(aiohttp_client: Any) -> None:
 @pytest.mark.parametrize("status", (101, 204, 304))
 @pytest.mark.parametrize("version", (HttpVersion10, HttpVersion11))
 async def test_no_body_for_1xx_204_304_responses(
-    aiohttp_client: Any, status: int, version: HttpVersion
+    aiohttp_client: AiohttpClient, status: int, version: HttpVersion
 ) -> None:
     """Test no body is present for for 1xx, 204, and 304 responses."""
 
-    async def handler(_):
+    async def handler(request: web.Request) -> web.Response:
         return web.Response(status=status, body=b"should not get to client")
 
     app = web.Application()
@@ -2164,3 +2269,47 @@ async def test_no_body_for_1xx_204_304_responses(
     assert TRANSFER_ENCODING not in resp.headers
     await resp.read() == b""
     resp.release()
+
+
+async def test_keepalive_race_condition(aiohttp_client: AiohttpClient) -> None:
+    protocol = None
+    orig_data_received = RequestHandler.data_received
+
+    def delay_received(self: RequestHandler[web.Request], data: bytes) -> None:
+        """Emulate race condition.
+
+        The keepalive callback needs to be called between data_received() and
+        when start() resumes from the waiter set within data_received().
+        """
+        orig_data_received(self, data)
+        if protocol is None:  # First request creating the keepalive connection.
+            return
+
+        assert self is protocol
+        assert protocol._keepalive_handle is not None
+        # Cancel existing callback that would run at some point in future.
+        protocol._keepalive_handle.cancel()
+        protocol._keepalive_handle = None
+
+        # Set next run time into the past and run callback manually.
+        protocol._next_keepalive_close_time = asyncio.get_running_loop().time() - 1
+        protocol._process_keepalive()
+
+    async def handler(request: web.Request) -> web.Response:
+        nonlocal protocol
+        protocol = request.protocol
+        return web.Response()
+
+    target = "aiohttp.web_protocol.RequestHandler.data_received"
+    with mock.patch(target, delay_received):
+        app = web.Application()
+        app.router.add_get("/", handler)
+        client = await aiohttp_client(app)
+
+        # Open connection, so we have a keepalive connection and reference to protocol.
+        async with client.get("/") as resp:
+            assert resp.status == 200
+        assert protocol is not None
+        # Make 2nd request which will hit the race condition.
+        async with client.get("/") as resp:
+            assert resp.status == 200
